Arreglando el models de el front, Agregamos que los stocks sean Doubles y un Boolean en Promocion para saber si esta activa o no

+++ BUEN_SABOR_BACKEND/src/main/java/com/lab4/buen_sabor_backend/model/SucursalInsumo.java
@@ -12,9 +12,9 @@ import java.util.*;
 @AllArgsConstructor
 public class SucursalInsumo extends Master {
 
-    private Integer stockMinimo;
-    private Integer stockActual;
-    private Integer stockMaximo;
+    private Double stockMinimo;
+    private Double stockActual;
+    private Double stockMaximo;
 
     @ManyToOne
     @JoinColumn(name = "sucursal_id")
